Rearranged the files structure
