--- src/blocks/modal.css
@@ -1,196 +1,0 @@
-.modal {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  background-color: rgba(0, 0, 0, 0.3);
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  right: 0;
-  z-index: 999;
-}
-
-.modal_type_successSignup {
-  height: 182px;
-}
-.modal__content {
-  width: 430px;
-  background-color: #ffffff;
-  box-shadow: 0px 0px 25px 0px #00000026;
-  border-radius: 16px;
-  position: absolute;
-  right: 50%;
-  top: 50%;
-  z-index: 999;
-  transform: translate(50%, -50%);
-}
-.modal__content .modal__close-button {
-  width: 40px;
-  height: 40px;
-  border: none;
-  background-color: transparent;
-  background-image: url("../images/back.png");
-  position: absolute;
-  top: -45px;
-  right: -45px;
-  z-index: 999;
-}
-.modal__content .modal__close-button:hover {
-  cursor: pointer;
-  opacity: 0.5;
-}
-.modal__title {
-  width: 358px;
-  height: 32px;
-  white-space: nowrap;
-  font-family: "Roboto", sans-serif;
-  font-style: Bold;
-  font-size: 24px;
-  font-weight: 900;
-  line-height: 28.13px;
-  text-align: left;
-  border: none;
-  background-color: transparent;
-  color: #1a1b22;
-  margin: 34px 35px 0;
-}
-.modal__form {
-  width: 358px;
-  height: auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  margin: 0 36px 0;
-}
-
-.modal__label {
-  width: 139px;
-  height: 14px;
-  display: flex;
-  flex-direction: column;
-  margin: 31px 0 31px;
-  font-family: "Inter", sans-serif;
-  font-size: 12px;
-  font-weight: 400;
-  line-height: 14.52px;
-  text-align: left;
-  color: #2f71e5;
-  border: none;
-  position: relative;
-}
-.modal__input {
-  width: 358px;
-  height: 14px;
-  white-space: nowrap;
-  border: none;
-  outline: none;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.2);
-  font-family: "Inter", sans-serif;
-  font-size: 14px;
-  font-weight: 400;
-  line-height: 16.94px;
-  text-align: left;
-  color: rgba(0, 0, 0, 0.2);
-  margin: 8px auto 11px 0;
-}
-.modal__input:valid {
-  border-bottom: 1px solid green;
-}
-.modal__input::placeholder {
-  color: #000000;
-  opacity: 0.2;
-}
-.modal__input_type_error {
-  border-bottom: 1px solid #ff0000;
-}
-.modal__error {
-  width: 173px;
-  height: 14px;
-  white-space: nowrap;
-  position: absolute;
-  top: 42px;
-  left: 30;
-  z-index: 3;
-  color: #ff0000;
-  font-family: "Inter", sans-serif;
-  font-size: 14px;
-  font-weight: 400;
-  line-height: 16.94px;
-  text-align: left;
-}
-.modal__error_visible {
-  opacity: 1;
-}
-.modal__submit-container {
-  width: 358px;
-  height: 100px;
-  display: flex;
-  flex-direction: column;
-  /* justify-content: center; */
-  margin: 39px 0 28px;
-}
-.modal__submit-button {
-  width: 360px;
-  height: 64px;
-  border-radius: 32px;
-  border: none;
-  outline: none;
-  background-color: #e6e8eb;
-  color: #b6bcbf;
-  font-family: "Roboto", sans-serif;
-  font-size: 18px;
-  font-weight: 500;
-  line-height: 24px;
-  text-align: center;
-  margin: 0 0 16px;
-  transition: background-color 0.3s ease;
-}
-
-.modal__submit-button_disabled {
-  background-color: #e6e8eb;
-  cursor: not-allowed;
-}
-.modal__submit-button_enabled {
-  background-color: #2f71e5;
-  box-shadow: 0px 5px 15px 0px #0e1a3933;
-  cursor: pointer;
-  opacity: 1;
-}
-
-.modal__option-button {
-  width: 226px;
-  height: 20px;
-  border: none;
-  background-color: transparent;
-  text-align: center;
-  margin: 0 auto 0;
-}
-.modal__or-text {
-  font-family: "Inter", sans-serif;
-  font-style: light;
-  font-size: 14px;
-  font-weight: 400;
-  line-height: 16.94px;
-  text-align: center;
-  padding-right: 4px;
-  color: #000000;
-}
-.modal__signin-text {
-  font-family: "Inter", sans-serif;
-  font-style: light;
-  font-size: 14px;
-  font-weight: 400;
-  line-height: 16.94px;
-  text-align: center;
-  color: #2f71e5;
-}
-.modal__signup-text {
-  font-family: "Inter", sans-serif;
-  font-size: 14px;
-  font-weight: 400;
-  line-height: 16.94px;
-  text-align: center;
-  color: #2f71e5;
-}
