+++ src/components/NewsCard/NewsCard.jsx
@@ -1,6 +1,6 @@
-import "../blocks/newsCard.css";
-import solidPattern from "../images/Group 12.png";
-import hollowPattern from "../images/Group 13.png";
+import "../NewsCard/newsCard.css";
+import solidPattern from '../../images/Group 12.svg';
+import hollowPattern from '../../images/Group 13.svg';
 
 const NewsCard = ({ card, onSave, onUnsave, currentUser }) => {
   if (!card) {
@@ -9,7 +9,6 @@ const NewsCard = ({ card, onSave, onUnsave, currentUser }) => {
 
   const isSaved =
     Array.isArray(card.saves) &&
-    currentUser &&
     card.saves.some((id) => id === currentUser?._id);
   const cardSaveButtonClassName = `card__save-button ${
     isSaved ? "card__save-button_saved" : "card__save-button"
@@ -26,7 +25,7 @@ const NewsCard = ({ card, onSave, onUnsave, currentUser }) => {
         <button
           type="button"
           className={cardSaveButtonClassName}
-          onClick={() => (isSaved ? onUnsave(card) : onSave(card))}
+          onClick={() => isSaved ? onUnsave(card) : onSave(card)}
         >
           <img
             src={isSaved ? solidPattern : hollowPattern}
@@ -38,7 +37,9 @@ const NewsCard = ({ card, onSave, onUnsave, currentUser }) => {
       </div>
       <div className="card__publishedAt">{card.publishedAt}</div>
       <div className="card__content">
-        <h2 className="card__title">{card.title || "No available title"}</h2>
+        <h2 className="card__title">
+          {card.title || "No available title"}
+        </h2>
         <p className="card__description">
           {card.description || "No available content"}
         </p>
