+++ src/components/NewsCard/newsCard.css
