+++ src/components/NewsSection/newsSection.css
