+++ src/components/SigninModal/SigninModal.jsx
@@ -1,40 +1,33 @@
-import ModalWithForm from "./ModalWithForm";
-import useForm from "../hooks/useForm";
-import "../blocks/modal.css";
+import ModalWithForm from "../ModalWithForm/ModalWithForm";
+import useForm from "../../hooks/useForm";
+import "../Modal/modal.css";
 
-const SignUpModal = ({
+const SigninModal = ({
   isOpen,
   isLoading,
   handleCloseModal,
-  handleSignup,
-  handleOpenSigninModal,
-  handleOpenSuccessSignupModal,
+  handleSignin,
+  handleOpenSignupModal,
 }) => {
+  console.log("is modal open:", isOpen);
+  if (!isOpen) return null;
+
   const { values, handleChange, errors, isFormValid } = useForm({
     email: "",
     password: "",
-    username: "",
   });
 
   const handleSubmit = (e) => {
     e.preventDefault();
-    handleSignup({
+    handleSignin({
       email: values.email,
       password: values.password,
-      username: values.username,
-    })
-      .then(() => {
-        handleCloseModal();
-        handleOpenSuccessSignupModal();
-      })
-      .catch((err) => {
-        console.log("Signup failed:", err);
-      });
+    });
   };
 
   return (
     <ModalWithForm
-      title="Sign up"
+      title="Sign in"
       onClose={handleCloseModal}
       isOpen={isOpen}
       onSubmit={handleSubmit}
@@ -49,7 +42,7 @@ const SignUpModal = ({
           value={values.email}
           onChange={handleChange}
           className="modal__input"
-          placeholder="Enter email"
+          placeholder="enter email"
           minLength="8"
           maxLength="40"
           aria-invalid={!!errors.email}
@@ -66,7 +59,7 @@ const SignUpModal = ({
           value={values.password}
           onChange={handleChange}
           className="modal__input"
-          placeholder="enter password"
+          placeholder="enter passwword"
           minLength="8"
           maxLength="40"
           aria-invalid={!!errors.password}
@@ -75,28 +68,8 @@ const SignUpModal = ({
           <span className="modal__error">{errors.password}</span>
         )}
       </label>
-      <label className="modal__label">
-        Username{" "}
-        <input
-          id="username"
-          required
-          name="username"
-          type="username"
-          value={values.username}
-          onChange={handleChange}
-          className="modal__input"
-          placeholder="enter your username"
-          minLength="2"
-          maxLength="40"
-          aria-invalid={!!errors.username}
-        />
-        {errors.username && (
-          <span className="modal__error">{errors.username}</span>
-        )}
-      </label>
       <div className="modal__submit-container">
         <button
-          type="submit"
           className={`modal__submit-button ${
             isFormValid
               ? "modal__submit-button_enabled"
@@ -104,19 +77,19 @@ const SignUpModal = ({
           } `}
           disabled={!isFormValid}
         >
-          {isLoading ? "Sign up..." : "Sign up"}
+          {isLoading ? "Sign in..." : "Sign in"}
         </button>
         <button
           type="button"
           className="modal__option-button"
-          onClick={handleOpenSigninModal}
+          onClick={handleOpenSignupModal}
         >
           <span className="modal__or-text">or</span>
-          <span className="modal__signup-text">Sign in</span>
+          <span className="modal__signin-text">Sign up</span>
         </button>
       </div>
     </ModalWithForm>
   );
 };
 
-export default SignUpModal;
+export default SigninModal;
